Types Improvement and Unit Tests (#3)
* int to NonNegativeFloat for amounts

* email type

* id as UUID type

* name constraints

* Password validation

* Organize types

* Some tests - needs refinement

* lint

* fix check_balance

* fix SecretStr type

* lint + fix

* test password validator + twiks

* rename tests

* test Name and StrongPassword types

* Reuse using fixture

* Redundant testing pydantic behaviour

* lint

---------

+++ app/services/balance_service.py
@@ -1,2 +1,2 @@
-async def check_balance(income: int, expenses: int) -> int:
+async def check_balance(income: float, expenses: float) -> float:
     return income - expenses
